memcached plugin: fix implicit conversion warning
  CC       src/memcached.lo
src/memcached.c:478:20: warning: implicit conversion loses integer precision: 'unsigned long' to 'int' [-Wshorten-64-to-32]
    int name_len = strlen(fields[1]);
        ~~~~~~~~   ^~~~~~~~~~~~~~~~~
1 warning generated.

+++ src/memcached.c
@@ -475,7 +475,7 @@ static int memcached_read(user_data_t *user_data) {
     if (strsplit(line, fields, 3) != 3)
       continue;
 
-    int name_len = strlen(fields[1]);
+    size_t name_len = strlen(fields[1]);
     if (name_len == 0)
       continue;
